+++ plone/app/dexterity/behaviors/exclfromnav.py
@@ -6,7 +6,36 @@ from plone.supermodel import model
 from z3c.form.interfaces import IAddForm
 from z3c.form.interfaces import IEditForm
 from zope import schema
+from zope.interface import implementer
+from zope.interface import Interface
 from zope.interface import provider
+from zope.schema.interfaces import IContextAwareDefaultFactory
+
+
+class IExcludeFromNavigationDefault(Interface):
+
+    def __call__():
+        """boolean if item is by default excluded from navigation or not.
+        """
+
+
+@implementer(IExcludeFromNavigationDefault)
+def default_exclude_false(context):
+    """provide a default adapter with the standard uses
+    """
+    return False
+
+
+@implementer(IExcludeFromNavigationDefault)
+def default_exclude_true(context):
+    """provide a alternative adapter with opposite default as standard
+    """
+    return True
+
+
+@provider(IContextAwareDefaultFactory)
+def default_exclude(context):
+    return IExcludeFromNavigationDefault(context)
 
 
 @provider(IFormFieldProvider)
@@ -27,10 +56,10 @@ class IExcludeFromNavigation(model.Schema):
         ),
         description=_(
             u'help_exclude_from_nav',
-            default=u'If selected, this item will not appear in the ' +
+            default=u'If selected, this item will not appear in the '
                     u'navigation tree'
         ),
-        default=False
+        defaultFactory=default_exclude,
     )
 
     directives.omitted('exclude_from_nav')